+++ lib/spior/reload.rb
@@ -1,0 +1,13 @@
+require 'tty-which'
+require_relative 'msg'
+
+module Spior
+  module Reload
+    def self.tor
+      if TTY::Which.exist?('systemctl')
+        system('sudo systemctl restart tor')
+        Msg.p "ip changed"
+      end
+    end
+  end
+end